Permite obtener employees de una company filtrados por status
